+++ tutorials/mct_model_garden/models_pytorch/yolov8/yolov8_postprocess.py
@@ -1,7 +1,8 @@
 from enum import Enum
 from typing import List, Tuple
 import numpy as np
-
+import cv2
+import matplotlib.pyplot as plt
 
 def nms(dets: np.ndarray, scores: np.ndarray, iou_thres: float = 0.5, max_out_dets: int = 300) -> List[int]:
     """
@@ -197,8 +198,6 @@ def scale_coords(kpts: np.ndarray, h_image: int, w_image: int, h_model: int, w_m
     # Scale and offset boxes
     kpts[..., 0] = (kpts[..., 0]  - deltaH) * scale_H
     kpts[..., 1] = (kpts[..., 1] - deltaW) * scale_W
-    # boxes[..., 2] = (boxes[..., 2] * H - deltaH) * scale_H
-    # boxes[..., 3] = (boxes[..., 3] * W - deltaW) * scale_W
 
     # Clip boxes
     kpts = clip_coords(kpts, h_image, w_image)
@@ -219,8 +218,6 @@ def clip_coords(kpts: np.ndarray, h: int, w: int) -> np.ndarray:
     """
     kpts[..., 0] = np.clip(kpts[..., 0], a_min=0, a_max=h)
     kpts[..., 1] = np.clip(kpts[..., 1], a_min=0, a_max=w)
-    # boxes[..., 2] = np.clip(boxes[..., 2], a_min=0, a_max=h)
-    # boxes[..., 3] = np.clip(boxes[..., 3], a_min=0, a_max=w)
     return kpts
 
 def postprocess_yolov8_detection(outputs: Tuple[np.ndarray, np.ndarray, np.ndarray],
@@ -406,3 +403,52 @@ def clip_coords(kpts: np.ndarray, h: int, w: int) -> np.ndarray:
     kpts[..., 0] = np.clip(kpts[..., 0], a_min=0, a_max=h)
     kpts[..., 1] = np.clip(kpts[..., 1], a_min=0, a_max=w)
     return kpts
+
+
+class COCODrawer:
+    def __init__(self, categories_file):
+        self.categories = self.get_categories(categories_file)
+
+    def get_categories(self, filename):
+        with open(filename, 'r') as f:
+            return [line.strip() for line in f.readlines()]
+
+    def draw_bounding_box(self, img, annotation, class_id, ax):
+        x_min, y_min = int(annotation[1]), int(annotation[0])
+        x_max, y_max = int(annotation[3]), int(annotation[2])
+        text = self.categories[int(class_id)]
+        cv2.rectangle(img, (x_min, y_min), (x_max, y_max), (0, 0, 255), 2)
+        ax.text(x_min, y_min, text, style='italic',
+                bbox={'facecolor': 'white', 'alpha': 0.7, 'pad': 5})
+
+    def draw_keypoints(self, img, keypoints):
+        skeleton = [
+            [0, 1], [0, 2], [1, 3], [2, 4],  # Head
+            [5, 6], [5, 7], [7, 9], [6, 8],  # Arms
+            [8, 10], [5, 11], [6, 12], [11, 12],  # Body
+            [11, 13], [12, 14], [13, 15], [14, 16]  # Legs
+        ]
+
+        # Draw skeleton lines
+        for connection in skeleton:
+            start_point = (int(keypoints[connection[0]][0]), int(keypoints[connection[0]][1]))
+            end_point = (int(keypoints[connection[1]][0]), int(keypoints[connection[1]][1]))
+            cv2.line(img, start_point, end_point, (255, 0, 0), 2)
+
+        # Draw keypoints as colored circles
+        for point in keypoints:
+            x, y = int(point[0]), int(point[1])
+            cv2.circle(img, (x, y), 3, (0, 255, 0), -1)
+    def annotate_image(self, img, b, s, c, k, scale, ax):
+        for index, row in enumerate(b):
+            if s[index] > 0.55:
+                self.draw_bounding_box(img, row * scale, c[index], ax)
+                if k is not None:
+                    self.draw_keypoints(img, k[index] * scale)
+
+    def plot_image(self, img, b, s, c, k=None):
+        fig, ax = plt.subplots()
+        self.annotate_image(img, b, s, c, k, scale=1, ax=ax)
+        ax.imshow(cv2.cvtColor(img, cv2.COLOR_BGR2RGB))
+        # plt.title(img.shape)
+        return fig, ax
